build: remove constant into individual app
