+++ packages/api-schema/src/index.ts
@@ -84,87 +84,3 @@ export type RegistrationFormData = {
   advisor: AdvisorFormData,
   team: TeamFormData,
 };
-
-export const RegistrationFormDataTemplate = {
-  students: [
-    {
-      drugAllergy: "",
-      quote: "",
-      foodAllergy: "",
-      email: "shinnapatjr@gmail.com",
-      grade: "Sophomore",
-      lineId: "shinnapat_krabphom",
-      disease: "",
-      middleNameEn: "",
-      middleNameTh: "",
-      firstnameEn: "Shinnapat",
-      firstnameTh: "ชินพรรธน์",
-      nickname: "เปปเปอร์",
-      phoneNumber: "0875908288",
-      foodType: "อะไรก็ได้",
-      prefixEn: "Mr.",
-      prefixTh: "นาย",
-      surnameEn: "Koparamestrisin",
-      surnameTh: "โกปาราเมศไตรสิน",
-    },
-    {
-      drugAllergy: "-",
-      quote: "-",
-      foodAllergy: "-",
-      email: "shin_gg@hotmail.com",
-      grade: "ม.7",
-      lineId: "มายไลน์ไอดี",
-      disease: "แพ้คนอย่างเธอ",
-      middleNameEn: "",
-      middleNameTh: "",
-      firstnameEn: "PPHamster",
-      firstnameTh: "พีพีแฮมสเตอร์",
-      nickname: "พีพี",
-      phoneNumber: "0812345678",
-      foodType: "บาร์บีก้อน",
-      prefixEn: "Mr.",
-      prefixTh: "นาย",
-      surnameEn: "Incursio",
-      surnameTh: "อินครูซิโอ้"
-    },
-    {
-      drugAllergy: "",
-      quote: "",
-      foodAllergy: "",
-      email: "",
-      grade: "",
-      lineId: "",
-      disease: "",
-      middleNameEn: "",
-      middleNameTh: "",
-      firstnameEn: "Third Person",
-      firstnameTh: "คนที่สาม",
-      nickname: "",
-      phoneNumber: "",
-      foodType: "",
-      prefixEn: "",
-      prefixTh: "",
-      surnameEn: "",
-      surnameTh: ""
-    },
-  ],
-  team: {
-    amount: 3,
-    school: "King Mongkut University of Technology Thonburi",
-    name: "Made In Abyss",
-    isComplete: true,
-  },
-  advisor: {
-    email: "",
-    lineId: "",
-    middleNameEn: "",
-    middleNameTh: "",
-    firstnameEn: "",
-    firstnameTh: "",
-    phoneNumber: "0812345678",
-    prefixEn: "",
-    prefixTh: "",
-    surnameEn: "",
-    surnameTh: ""
-  }
-};
